MyLinkedList class that implements MyList and contains methods of LinkedList elements
defined a private inner class Node

head references the first node in the list.
tail references the last node in the list
size keep track of the number of elements in the list

add method creates new Node class element, that contains given element and dont have next element(null) and the tail;
if our LinkedList is empty, isEmpty method returns true, and we say that head equals to newNode. If its not: we say that next tail is newNode. Then we make our size +1;