MyLinkedList class that implements MyList and contains methods of LinkedList elements
defined a private inner class Node

head references the first node in the list.
tail references the last node in the list
size keep track of the number of elements in the list

add method creates new Node class element, that contains given element and dont have next element(null) and the tail;
if our LinkedList is empty, isEmpty method returns true, and we say that head equals to newNode. If its not: we say that next tail is newNode. Then we make our size +1;

ill be sure, this method was hard to me. So, at first we check if index is in range of list. then we create Node "currentNode" and if index is less than half of size, we say that it equals to head; else we consider currentNode as tail;

contains method check if given object equals to one of elements of list.